requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
click==8.1.7
cmake==3.29.2
contourpy==1.1.1
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cycler==0.12.1
decorator==5.1.1
Django==4.2.11
django-crispy-forms==2.1
dlib==19.9.0
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fonttools==4.51.0
imageio==2.34.1
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.7.5
matplotlib-inline==0.1.7
networkx==3.1
numpy==1.24.4
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
threadpoolctl==3.5.0
tifffile==2023.7.10
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
zipp==3.18.1
